requirements.txt
# ─── LangChain ────────────────────────────────────────────────────────────────
langchain==0.3.14
langchain-core==0.3.33
langchain-community==0.3.14
langchain-groq==0.2.4
langchain-huggingface==0.1.2
langchain-chroma==0.2.1

# ─── AI & Embeddings ──────────────────────────────────────────────────────────
groq==0.13.1
sentence-transformers==3.4.1
chromadb==0.6.3

# ─── Document Loaders ─────────────────────────────────────────────────────────
pypdf==5.1.0
docx2txt==0.8
unstructured==0.16.23

# ─── Voice ────────────────────────────────────────────────────────────────────
streamlit-audiorecorder==0.0.5
pydub==0.25.1
gTTS==2.5.3

# ─── UI ───────────────────────────────────────────────────────────────────────
streamlit==1.42.0
python-dotenv==1.0.1
